% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tangfunction.R
\name{tan_agreg}
\alias{tan_agreg}
\title{Agrégation des données}
\usage{
tan_agreg(tout, mois)
}
\arguments{
\item{tout}{Si vous souhaitez afficher toutes les dépenses, indiquez TRUE}

\item{mois}{Si vous souhaitez sélectionner un mois précis, faites-le ici à l'aide du numéro du mois (ex: 05)}
}
\value{
Un tibble qui affiche l'agrégation souhaitée
}
\description{
Cette fonction permet d'afficher une agrégation de vos dépenses par mois et par année.
}
