% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tangfunction.R
\name{tan_graphique}
\alias{tan_graphique}
\title{Représentation graphique des dépenses mensuelles.}
\usage{
tan_graphique(data, tout = TRUE, annee)
}
\arguments{
\item{data}{Votre jeu de donnée importé par tangimport}

\item{tout}{Si vous souhaitez afficher les dépenses de toute la période, indiquez TRUE}

\item{annee}{Si vous souhaitez sélectionner une année précise, indiquez là ici en format numérique (ex: 2020)}
}
\value{
Un graphique avec l'évolution mensuelle de vos dépenses sur 12 mois
}
\description{
Cette fonction permet d'afficher un graphique de vos dépenses mensuelles sur 12 mois, avec une couleur
différente par année
}
